+++ lib/params/arrowparams.cpp
@@ -115,7 +115,7 @@ Validate(bool doOverride){
 	//Set the vector length so that the max arrow is 10% of the larger of the x or y size of scene
 	//Check the rake extents.  If doOverride is true, set the extents to the bottom of the data domain. If not, 
 	//shrink the extents to fit inside the domain.
-	const float* extents = ds->getLocalExtents();
+	const double* extents = ds->getLocalExtents();
 	vector<double>newExtents(3,0.);
 	if (doOverride) {
 		for (int i = 0; i<3; i++){
@@ -326,7 +326,7 @@ double ArrowParams::calcDefaultScale(){
 	string varname;
 	double maxvarvals[3];
 	DataStatus* ds = DataStatus::getInstance();
-	const float* stretch = ds->getStretchFactors();
+	const double* stretch = ds->getStretchFactors();
 	for (int i = 0; i<3; i++){
 		varname = GetFieldVariableName(i);
 		if (varname == "0" || !ds->getDataMgr()) maxvarvals[i] = 0.;
@@ -335,7 +335,7 @@ double ArrowParams::calcDefaultScale(){
 		}
 	}
 	for (int i = 0; i<3; i++) maxvarvals[i] *= stretch[i];
-	const float* extents = DataStatus::getInstance()->getLocalExtents();
+	const double* extents = DataStatus::getInstance()->getLocalExtents();
 	double maxVecLength = (double)Max(extents[3]-extents[0],extents[4]-extents[1])*0.1;
 	double maxVecVal = Max(maxvarvals[0],Max(maxvarvals[1],maxvarvals[2]));
 	if (maxVecVal == 0.) return(maxVecLength);
